First push of branch. Does not compile.
The getInfo() method in View does not override the getInfo() method in BasicView. In ServerInfo(), this line     this.version = Uuid.parse(SERVER_VERSION);
is throwing an IOException and needs a try-catch block. However, when I did that, the version variable was not properly declared and there was still an error.